requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
streamlit>=1.25.0
plotly>=5.15.0
schedule>=1.2.0
websockets>=11.0
aiohttp>=3.8.0
cryptography>=41.0.0
pycryptodome>=3.18.0
holidays>=0.39
pytz>=2023.3
python-telegram-bot>=21.0.0
